Cast to int to avoid gcc complaint from python enum registration.

+++ python/rpmmodule.c
@@ -459,7 +459,7 @@ void init_rpm(void)
 
     REGISTER_ENUM(RPMDBI_PACKAGES);
 
-    REGISTER_ENUM(RPMAL_NOMATCH);
+    REGISTER_ENUM((int)RPMAL_NOMATCH);
 }
 
 /*@}*/
